Refactor commands into their own file
