+++ leveler/comands/badge.py
@@ -2,10 +2,10 @@ from redbot.core import commands
 import discord
 from redbot.core.utils.chat_formatting import pagify
 import operator
-from .leveler import db
+from leveler.leveler import db
 from redbot.core import checks, bank
-from .path_munger import fileIO
-from .static_methods import _badge_convert_dict, _is_hex
+from ..path_munger import fileIO
+from ..static_methods import _badge_convert_dict, _is_hex
 
 prefix = "!"
